+++ src/app/sass/components/_member.scss
@@ -1,8 +1,22 @@
 .about {
   &__member {
+    width: 100%;
+    height: 100%;
     display: flex;
     flex-direction: column;
     gap: 1rem;
+    border-radius: .5rem .5rem 0 0;
+    box-shadow: rgba(0, 0, 0, 0.16) 0px .1rem .3rem;
+    padding: 1rem;
+    cursor: pointer;
+
+    // &__member:hover > .front{
+    //   transform: perspective(600px) rotateY(-180deg);
+    // }
+
+    // &__member:hover > .back{
+    //   transform: perspective(600px) rotateY(0deg);
+    // }
 
     &-img {
       width: 20vw;
@@ -40,5 +54,58 @@
       height: 4rem;
       margin-left: -1rem;
     }
+
+    &-button {
+      width: fit-content;
+      border: 1px solid $color-shadow !important;
+    }
   }
+}
+
+.center {
+  // position: relative;
+  // top: 500%;
+  // left: 20%;
+  // transform: translate(-50%, -50%);
+}
+
+.front,
+.back {
+  width: 100%;
+  height: 100%;
+  // overflow: hidden;
+  backface-visibility: hidden;
+  transition: transform .6s linear;
+  position: relative;
+}
+
+.back {
+  position: absolute;
+  top: -.1%;
+  left: -.2%;
+  border-radius: .5rem .5rem 0 0;
+  box-shadow: rgba(0, 0, 0, 0.16) 0px .1rem .3rem;
+  padding: 1rem;
+}
+
+.front{
+  transform: perspective(600px) rotateY(0deg);
+}
+
+.back{
+  transform: perspective(600px) rotateY(180deg);
+  background-color: #f1f1f1;
+}
+
+.back-content{
+  width: 100%;
+  text-align: center;
+  color: #333333;
+}
+
+.card:hover > .front{
+  transform: perspective(600px) rotateY(-180deg);
+}
+.card:hover > .back{
+  transform: perspective(600px) rotateY(0deg);
 }
