+++ src/app/sass/pages/_about-us.scss
@@ -3,6 +3,7 @@
   &__title {
     text-align: center;
     font-size: 2rem;
+    padding-bottom: 3rem;
   }
 
   &__description {
